+++ lib/ole_qa_framework/COMPATIBILITY.rb
@@ -15,6 +15,6 @@
 module OLE_QA
   module Framework
     # The most recent version of OLE with which this framework was tested for compatibility.
-    OLE_VERSION = '1.5.0 - M2 - r18595 - 2014/04/17'
+    OLE_VERSION = '1.5.0 - M2 - r19440 - 2014/06/23'
   end
 end